refactor tests/ code to reduce duplicate code
