+++ tests/utils.py
@@ -1,52 +1,68 @@
-# utils.py: useful utils for mocking requests and responses for testing
-from api_app.script_analyzers.observable_analyzers import pulsedive
+import hashlib
+from abc import ABCMeta
+from unittest.mock import patch
+
+from api_app.models import Job
+from api_app.script_analyzers.observable_analyzers import (
+    pulsedive,
+    otx,
+    vt2_get,
+    vt3_get,
+    misp,
+    onyphe,
+    ha_get,
+    thug_url,
+    urlhaus,
+    googlesf,
+    fortiguard,
+)
+
+from .mock_utils import mock_connections, mocked_requests_noop
 
 
-# class for mocking responses
-class MockResponse:
-    def __init__(self, json_data, status_code):
-        self.json_data = json_data
-        self.status_code = status_code
-        self.text = ""
-        self.content = b""
-
-    def json(self):
-        return self.json_data
-
-    def raise_for_status(self):
-        pass
-
-
-# a mock response class that has no operation
-class MockResponseNoOp:
-    def __init__(self, json_data, status_code):
-        pass
-
-    def search(self, **kwargs):
-        return {}
-
-    def query(self, val):
-        return {}
-
-
-def mocked_requests(*args, **kwargs):
-    return MockResponse({}, 200)
-
-
-def mocked_docker_analyzer_get(*args, **kwargs):
-    return MockResponse(
-        {"key": "test", "returncode": 0, "report": {"test": "This is a test."}}, 200
-    )
-
-
-def mocked_docker_analyzer_post(*args, **kwargs):
-    return MockResponse({"key": "test", "status": "running"}, 202)
-
-
-class CommonTestCases:
+# Abstract Base classes constructed for most common occuring combinations
+# to avoid duplication of code
+class CommonTestCases_observables(metaclass=ABCMeta):
     """
-    Tests which are common for all types of observables.
+    Includes tests which are common for all types of observables.
     """
+
+    def setUp(self):
+        params = self.get_params()
+        params["md5"] = hashlib.md5(
+            params["observable_name"].encode("utf-8")
+        ).hexdigest()
+        test_job = Job(**params)
+        test_job.save()
+        self.job_id = test_job.id
+        self.observable_name = test_job.observable_name
+        self.observable_classification = test_job.observable_classification
+
+    def test_vt3_get(self, mock_get=None, mock_post=None):
+        report = vt3_get.VirusTotalv3(
+            "VT_v3_Get",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+    def test_vt2_get(self, mock_get=None, mock_post=None):
+        report = vt2_get.VirusTotalv2(
+            "VT_v2_Get",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+    def test_otx(self, mock_get=None, mock_post=None):
+        report = otx.OTX(
+            "OTX", self.job_id, self.observable_name, self.observable_classification, {}
+        ).start()
+        self.assertEqual(report.get("success", False), True)
 
     def test_pulsevide(self, mock_get=None, mock_post=None):
         report = pulsedive.Pulsedive(
@@ -55,5 +71,95 @@ class CommonTestCases:
             self.observable_name,
             self.observable_classification,
             {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+    @mock_connections(patch("pymisp.ExpandedPyMISP", side_effect=mocked_requests_noop))
+    def test_misp_first(self, *args):
+        report = misp.MISP(
+            "MISP_FIRST",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {"api_key_name": "FIRST_MISP_API", "url_key_name": "FIRST_MISP_URL"},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+
+class CommonTestCases_ip_url_domain(metaclass=ABCMeta):
+    """
+    Tests which are common for IP, URL, domain types.
+    """
+
+    def test_gsf(self, mock_get=None, mock_post=None):
+        report = googlesf.GoogleSF(
+            "GoogleSafeBrowsing",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+    def test_onyphe(self, mock_get=None, mock_post=None):
+        report = onyphe.Onyphe(
+            "ONYPHE",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+
+class CommonTestCases_ip_domain_hash(metaclass=ABCMeta):
+    """
+    Tests which are common for IP, domain, hash types.
+    """
+
+    def test_ha_get(self, mock_get=None, mock_post=None):
+        report = ha_get.HybridAnalysisGet(
+            "HybridAnalysis_Get_Observable",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+
+class CommonTestCases_url_domain(metaclass=ABCMeta):
+    """
+    Tests which are common for URL and Domain types.
+    """
+
+    def test_fortiguard(self, mock_get=None, mock_post=None):
+        report = fortiguard.Fortiguard(
+            "Fortiguard",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+    def test_urlhaus(self, mock_get=None, mock_post=None):
+        report = urlhaus.URLHaus(
+            "URLhaus",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            {},
+        ).start()
+        self.assertEqual(report.get("success", False), True)
+
+    def test_thug_url(self, mock_get=None, mock_post=None):
+        additional_params = {"test": True}
+        report = thug_url.ThugUrl(
+            "Thug_URL_Info",
+            self.job_id,
+            self.observable_name,
+            self.observable_classification,
+            additional_params,
         ).start()
         self.assertEqual(report.get("success", False), True)
